upd: pigo の検出器いれた
