+++ go.mod
@@ -4,15 +4,13 @@ go 1.22.0
 
 toolchain go1.22.5
 
-require (
-	github.com/esimov/pigo v1.4.6
-	github.com/hajimehoshi/ebiten/v2 v2.8.5
-)
+require github.com/hajimehoshi/ebiten/v2 v2.8.5
 
 require (
 	github.com/ebitengine/gomobile v0.0.0-20241016134836-cc2e38a7c0ee // indirect
 	github.com/ebitengine/hideconsole v1.0.0 // indirect
 	github.com/ebitengine/purego v0.8.1 // indirect
+	github.com/esimov/pigo v1.4.6 // indirect
 	github.com/jezek/xgb v1.1.1 // indirect
 	golang.org/x/sync v0.10.0 // indirect
 	golang.org/x/sys v0.28.0 // indirect
